requirements.txt
altgraph==0.17.2
appdirs==1.4.4
arrow==1.2.0
auto-py-to-exe==2.10.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
DateTime==4.3
distlib==0.3.3
Eel==0.12.4
fake-useragent==0.1.11
filelock==3.3.0
Flask==2.0.2
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
greenlet==1.1.2
gspread==4.0.1
idna==2.7
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
parse==1.19.0
pefile==2021.9.3
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.4.0
py==1.10.0
py2exe==0.10.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.2.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyppeteer==0.2.6
pyquery==1.4.3
pyTelegramBotAPI==4.1.0
pytest==3.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
pywin32-ctypes==0.2.0
pyzbar==0.1.8
qrcode==7.3.1
redis==3.5.3
requests==2.20.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
telegram==0.0.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.24.3
virtualenv==20.8.1
virtualenv-clone==0.5.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
whichcraft==0.6.1
xvfbwrapper==0.2.9
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
